requirements.txt
plyer==2.1.0
pypiwin32==223
pywin32==306
speedtest-cli==2.1.3